moved db functions into individual files and changed some useEffect logic
