+++ src/pages/Microbots.jsx
@@ -1,106 +1,14 @@
-import supabase from '../config/supabaseClient';
-import { useEffect, useState } from 'react';
+import { useState, useEffect } from 'react'
+import { fetchMembers } from '../utils/fetchMembers'
 
 import BattlebotCard from '../components/battlebotCard/battlebotCard';
 
-function GetEmail({ email, position }) {
-    if (position === 'President') {
-        return <p>Contact our president at {email}</p>;
-    }
-}
-
-function ToggleSignup({ offSeason, formLink }) {
-    if (offSeason) {
-        return <h2 id="SignUpUnavailable">SIGN-UP UNAVAILABLE</h2>;
-    } else {
-        return (
-            <h2>
-                <a href={formLink} target="_blank" rel="noreferrer" id="SignUp">
-                    SIGN UP TODAY
-                </a>
-            </h2>
-        );
-    }
-}
-
-function ThisSeason({ Battlebots, currentSemester }) {
-    /* {fetchError && (<p>{fetchError}</p>)}
-	{Battlebots && (
-		if(currentSemester != "Spring 2023") {
-			return(
-				<p>No competitors have signed up yet. Stay tuned!</p>
-			);
-		}
-		else {
-			return(
-				<div className = "battlebotCardContainer">
-					{Battlebots.map(battlebot => (
-						<BattlebotCard key={battlebot.id} battlebot={battlebot} currentSemester={currentSemester}/>
-					))}
-				</div>
-			);
-		}
-	)} */
-}
-
-function PrevSemesterButtons() {
-    const prevSemesters = ['Fall 2022', 'Spring 2023'];
-
-    const list = prevSemesters.map((semester) => (
-        <div key={semester}>
-            <div className="battlebotSeasonCard">
-                <h3>{semester}</h3>
-            </div>
-        </div>
-    ));
-
-    return <>{list}</>;
-}
-
 const Microbots = () => {
-    const [fetchError, setFetchError] = useState(null);
-    const [Battlebots, setBattlebots] = useState(null);
-    const [Members, setMembers] = useState([]);
-
-    ////////////////////////////////////////////////////
-    //Manually change these as needed
-    //Can't think of a good way to do it automatically
+    
     const currentSemester = 'Fall 2023';
     const offSeason = true;
     const formLink = 'https://www.youtube.com/@Ayden3D';
-    ////////////////////////////////////////////////////
 
-    useEffect(() => {
-        const fetchBattlebots = async () => {
-            const { data, error } = await supabase.from('Battlebots').select();
-
-            if (error) {
-                setFetchError('Error: failed to fetch battlebot data');
-                console.log(error);
-                setBattlebots(null);
-            }
-            if (data) {
-                setBattlebots(data);
-                setFetchError(null);
-            }
-        };
-        fetchBattlebots();
-
-        const fetchMembers = async () => {
-            const { data, error } = await supabase.from('Members').select();
-
-            if (error) {
-                setFetchError('Error: failed to fetch member data');
-                console.log(error);
-                setMembers(null);
-            }
-            if (data) {
-                setMembers(data);
-                setFetchError(null);
-            }
-        };
-        fetchMembers();
-    }, []);
 
     return (
         <div className="page microbots">
@@ -108,7 +16,7 @@ const Microbots = () => {
                 <h1>Microbots</h1>
                 <p>
                     As part of RIOT's vision, every year, we host a microbots competition where students new to robotics
-                    pilot small robots in a variety of different competitive events. We’ve hosted each previous semester
+                    pilot small robots in a variety of different competitive events. We've hosted each previous semester
                     two separate battle bot tournaments where students build a plastic, 3D printed, remote controlled
                     battle bot in a competition to fight and survive against another designs. Our winners receive
                     special awards as the crowned champion of the microbots tournament each time a competition is
@@ -128,7 +36,7 @@ const Microbots = () => {
             </center>
             <div>
                 <center>
-                    <ToggleSignup key={offSeason} offSeason={offSeason} formLink={formLink} />
+                    {/*<ToggleSignup key={offSeason} offSeason={offSeason} formLink={formLink} />*/}
                 </center>
             </div>
             <center>
@@ -141,33 +49,8 @@ const Microbots = () => {
                 </div>
                 <div>
                     <h2>Questions?</h2>
-                    {fetchError && <p>{fetchError}</p>}
-                    {Members.map((member) => (
-                        <GetEmail key={member.id} email={member.email} position={member.position} />
-                    ))}
                 </div>
             </center>
-
-            <div>
-                <center>
-                    {/* <h1>This Season</h1> */}
-                    {/* <ThisSeason key={Battlebots.id} Battlebots={Battlebots} currentSemester={currentSemester}/> */}
-                    {/* {fetchError && (<p>{fetchError}</p>)}
-				{Battlebots && (
-					<div className = "battlebotCardContainer">
-						{Battlebots.map(battlebot => (
-							<ThisSeason key={battlebot.id} battlebot={battlebot} currentSemester={currentSemester}/>
-						))}
-					</div>
-				)} */}
-                </center>
-            </div>
-            {/* <div id = "prevSemesters">
-				<h1>Previous Seasons</h1>
-				<div className = "battlebotSeasonCardContainer">
-					<PrevSemesterButtons/>
-				</div>
-			</div> */}
         </div>
     );
 };
