+++ src/styles/index.css
@@ -789,14 +789,6 @@ div.desc {
 /*---------------------------------------*/
 /*------------BATTLEBOT CARDS------------*/
 /*---------------------------------------*/
-.battlebotCard {
-    box-shadow: var(--lightShadow);
-    width: 300px;
-    height: 125px;
-    padding: 5px 20px 0px 20px;
-    margin: 3%;
-    text-align: center;
-}
 
 .battlebotCardContainer {
     display: grid;
